moved abstracts and fixed more lint errors
