+++ pages/abstracts/AbstractManagementView.page.ts
