+++ pages/admin/organisationen/KlasseCreationView.neu.page.ts
@@ -1,6 +1,6 @@
 import { type Locator, Page } from '@playwright/test';
 import { Autocomplete } from '../../../elements/Autocomplete';
-import { AbstractAdminPage } from '../../AbstractAdminPage.page';
+import { AbstractAdminPage } from '../../abstracts/AbstractAdminPage.page';
 
 export interface KlasseCreationParams {
     schulname: string,
