+++ pages/admin/organisationen/KlasseManagementView.neu.page.ts
@@ -2,7 +2,7 @@ import { expect, type Locator, Page } from '@playwright/test';
 import { Autocomplete } from '../../../elements/Autocomplete';
 import { DataTable } from '../../components/DataTable.page';
 import { KlasseDetailsViewPage } from './KlasseDetailsView.neu.page';
-import { AbstractManagementViewPage } from '../../AbstractManagementView.page';
+import { AbstractManagementViewPage } from '../../abstracts/AbstractManagementView.page';
 
 export class KlasseManagementViewPage extends AbstractManagementViewPage {
   /* add global locators here */
